requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
asgiref==3.8.1
asttokens==2.4.1
autopep8==2.3.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
billiard==4.2.0
build==1.2.1
CacheControl==0.14.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comtypes==1.4.5
crashtest==0.4.1
crispy-bootstrap5==2024.2
cron-descriptor==1.4.3
cryptography==42.0.8
decorator==5.1.1
defusedxml==0.8.0rc2
distlib==0.3.8
distro==1.9.0
Django==5.0.7
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-crispy-forms==2.2
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-filter==24.2
django-templated-mail==1.1.1
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
dulwich==0.21.7
executing==2.0.1
Faker==26.0.0
fastjsonschema==2.20.0
filelock==3.15.4
flower==2.0.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanize==4.10.0
idna==3.7
installer==0.7.0
ipython==8.26.0
jaraco.classes==3.4.0
jedi==0.19.1
joblib==1.4.2
keyring==24.3.1
kombu==5.3.7
Markdown==3.6
matplotlib-inline==0.1.7
more-itertools==10.3.0
msgpack==1.0.8
numpy==2.0.1
oauthlib==3.2.2
openai==1.37.1
packaging==24.1
Panda3D==1.10.14
panda3d-gltf==1.2.0
panda3d-simplepbr==0.12.0
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
pkginfo==1.11.1
platformdirs==4.2.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
PyAudio==0.2.14
pycodestyle==2.12.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pygame==2.6.0
Pygments==2.18.0
PyJWT==2.8.0
pyperclip==1.9.0
pypiwin32==223
pyproject_hooks==1.1.0
pyTelegramBotAPI==4.21.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.4
python3-openid==3.2.0
pyttsx3==2.90
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
rapidfuzz==3.9.4
redis==5.0.7
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
scikit-learn==1.5.1
scipy==1.14.0
screeninfo==0.8.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
soupsieve==2.5
SpeechRecognition==3.10.4
SQLAlchemy==2.0.31
sqlparse==0.5.0
srt==3.5.3
stack-data==0.6.3
threadpoolctl==3.5.0
tomlkit==0.12.5
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
trove-classifiers==2024.7.2
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
ursina==7.0.0
vine==5.1.0
virtualenv==20.26.3
vosk==0.3.45
wcwidth==0.2.13
websockets==12.0
